requirements.txt
aiofiles==23.2.1
aiogram==3.2.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
bibtexparser==2.0.0b4
certifi==2023.11.17
charset-normalizer==3.3.2
clldutils==3.22.1
colorama==0.4.6
colorlog==6.8.0
csvw==3.2.1
dlinfo==1.2.1
ffmpeg==1.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.2
future==0.18.3
huggingface-hub==0.20.1
idna==3.6
isodate==0.6.1
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
language-tags==1.2.0
lxml==4.9.4
magic-filter==1.0.12
Markdown==3.5.1
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.2
packaging==23.2
phonemizer==3.2.1
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pylatexenc==2.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.32.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rpds-py==0.15.2
rsa==4.9
safetensors==0.4.1
segments==2.2.1
six==1.16.0
tabulate==0.9.0
Telethon==1.33.1
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
